% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcFocalOffset.R
\name{CalcFocalOffset}
\alias{CalcFocalOffset}
\title{Calculate the spatial offset using a focal mean}
\usage{
CalcFocalOffset(covStack, r, spatGrid)
}
\arguments{
\item{covStack}{a RasterStack object with each layer containing annual
sagebrush percent cover data}

\item{r}{radius to use for the circular focal mean.  Must be in same
linear units as the covStack CRS}

\item{spatGrid}{a SpatialPolygonsDataFrame to which values for the spatial
offset will be appended as new columns}
}
\value{
a list object with the overall equilibrium cover and the output
    SpatialPolygonsDataFrame
}
\description{
Calculates a temporal average of all input rasters, then a focal mean using
circular moving window with a given radius. The spatial offset is calculated
as the difference of the log(focal mean) - log(equilibrium cover).
}
\examples{
# Make a cover raster stack
library(raster)
rasterDir <- file.path("L:/102 - USGS/102 - 88 - SageCastWY/Data/BackInTime/Sagebrush/WYGrouseCoreAreas/North Laramie")
covRasts <- list.files(rasterDir, "*.tif", full.names = T, ignore.case = T)
covStack <- raster::stack(covRasts)
spatGrid <- readRDS("L:/102 - USGS/102 - 88 - SageCastWY/Code/GridExample.RDS")
# Calc focal offset
focalOffset <- CalcFocalOffset(covStack, r = 800, spatGrid = spatGrid) # takes about 5 minutes to run
focalOffset$equalibriumCover
summary(focalOffset$offsetData)
}
