% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSpatialSummaries.R
\name{CalcSpatialSummaries}
\alias{CalcSpatialSummaries}
\title{Calculate summaries of the posterior predictive distribution of proportion
of cells that exceed a target value.}
\usage{
CalcSpatialSummaries(cover, target)
}
\arguments{
\item{cover}{An array of cover simulations from \code{\link{SimDAST}}.}

\item{target}{The target value of percent cover.}
}
\value{
A dataframe.
}
\description{
`CalcSpatialSummaries()` computes summary statistics of the posterior
predictive distribution of the proportion of cells that exceed a
target value.
}
