% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTemporalSummaries.R
\name{CalcTemporalSummaries}
\alias{CalcTemporalSummaries}
\title{Calculate summaries of the posterior predictive distribution
of average cover.}
\usage{
CalcTemporalSummaries(cover)
}
\arguments{
\item{cover}{An array of cover simulations from \code{\link{SimDAST}}.}
}
\value{
A dataframe.
}
\description{
`CalcTemporalSummaries()` computes summary statistics of the posterior
predictive distribution of average cover in each year. Cover is averaged
over space and then the summaries are computed over MCMC iterations.
}
