% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeBayesPValues.R
\name{ComputeBayesPValues}
\alias{ComputeBayesPValues}
\title{Compute Bayesian p-values from posterior predictive checks}
\usage{
ComputeBayesPValues(df, theta)
}
\value{
A data frame.
}
\description{
Compute test stastics from the posterior predictive distribution
and calculate Bayesian p-values for posterior predictive checks of
model lack of fit. A Chi-square test statistic is used here.
}
