% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetData}
\alias{GetData}
\title{Read in and format data}
\usage{
GetData(coverPath, weatherPath, resFactor = NULL, totalToRead = NULL)
}
\arguments{
\item{coverPath}{A character string specifying the path to the cover data
of interest. Assumes tif files from the USGS BIT data.}

\item{weatherPath}{A character string specifying the path to the weather
data of interest. Assumes tif files from PRISM.}

\item{resFactor}{An integer for the factor by which to decrease
the resolution of the spatial data. Default is \code{NULL}, implying
no change to the native resolution.}

\item{totalToRead}{An integer for the number of files (years) to read.
Default is \code{NULL}, meaning that all years are read in and
collated.}
}
\value{
A large list.
}
\description{
Creates a list of geospatial data for modeling
}
