% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataSample.R
\name{GetDataSample}
\alias{GetDataSample}
\title{Takes a spatially balanced random sample from the full grid of data
to generate a subset of cells to use for modeling fitting.}
\usage{
GetDataSample(cellIds, proportionToFit, spatialDf)
}
\arguments{
\item{cellIds}{A numeric vector of cell ids.}

\item{proportionToFit}{The proportion of total cells to sample.}

\item{spatialDf}{A spatial points polygon to be sampled.}
}
\value{
A list.
}
\description{
Takes a spatially balanced random sample from the full grid of data
to generate a subset of cells to use for modeling fitting.
}
