% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInitialConditions.R
\name{GetInitialConditions}
\alias{GetInitialConditions}
\title{Extracts initial conditions for equilibrium simulations and projections.}
\usage{
GetInitialConditions(df)
}
\arguments{
\item{df}{A data frame of counts (cover) by Cell and Year.}
}
\value{
A data frame.
}
\description{
Extracts initial conditions for equilibrium simulations and projections.
}
