% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetKnots.R
\name{GetKnots}
\alias{GetKnots}
\title{Generate knots}
\usage{
GetKnots(sfObj, spatialRange)
}
\arguments{
\item{sfObj}{A prototypical \code{sf} object for the spatial domain.}

\item{spatialRange}{A list of results returned by
\code{GetSpatialDependence}.}
}
\value{
A large list.
}
\description{
Creates a K matrix and knot placements for the spatial
   autoregressive model.
}
