% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPriors.R
\name{GetPriors}
\alias{GetPriors}
\title{Handy function to return priors from Tredennick et al. 2016}
\usage{
GetPriors()
}
\value{
A data frame.
}
\description{
Reads in MCMC results from Tredennick et al. 2016 (Ecosphere)
    and computes summary statistics for prior distributions
    of the intercept and density-dependence parameters.
}
