% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpatialDependence.R
\name{GetSpatialDepdendence}
\alias{GetSpatialDepdendence}
\title{Estimate spatial autocorrelation}
\usage{
GetSpatialDepdendence(dataList, model)
}
\arguments{
\item{dataList}{A list of the data, including counts and the
list of spatially gridded covariates.}

\item{model}{The poisson regression model to fit.}
}
\value{
A list.
}
\description{
Uses a variogram to estimate the distance at which spatial
   autocorrelation goes away.
}
