% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPosteriors.R
\name{PlotPosteriors}
\alias{PlotPosteriors}
\title{Plots and processes posterior distributions with prior overlaid.}
\usage{
PlotPosteriors(
  coreId,
  coreName,
  filePath = "./Output/MCMC/",
  figPath = "./Figures/",
  postPath = "./Output/Posteriors/",
  params = NULL,
  paramLabs = NULL
)
}
\arguments{
\item{coreId}{A string for the core id.}

\item{coreName}{A string for the core name.}

\item{filePath}{A string specifying the file path with MCMC subdirs.}

\item{figPath}{A string specifying the file path with figure subdirs.}

\item{postPath}{A string specifying the file path with Posteriors subdirs.}

\item{params}{A character vector with parameter names to extract from
the Stan MCMC results. Must match parameter names in the mcmc object.}

\item{paramLabs}{A factor vector specifying the order and names for each
parameter to be plotted.}
}
\value{
Nothing. Saves a list of ggplot object to specified directory.
}
\description{
Plots and processes posterior distributions with prior overlaid.
}
