% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSensitivityTarget.R
\name{RunSensitivityTarget}
\alias{RunSensitivityTarget}
\title{Performs sensitivity to climate analysis for fitted model with spatial
   component to test against targets.}
\usage{
RunSensitivityTarget(posteriors, eta, delta, coreName, target)
}
\arguments{
\item{posteriors}{A k*p matrix, where k is the number of MCMC iterations
and p is the number of parameters. Column names must be: \code{Beta[1]},
\code{Beta[2]}, \code{Beta[3]}, \code{Beta[4]}, \code{sigma_a}, and
\code{sigma_y}.}

\item{eta}{A k*c matrix, where k is the number of MCMC iterations and
c is the number of grid cells in the spatial domain.}

\item{delta}{The deviation from mean climate. A named vector: one element for
"precipitation" and one element for "temperature".}

\item{coreName}{TBD}

\item{target}{A percent cover target value.}
}
\value{
A dataframe.
}
\description{
Performs sensitivity to climate analysis for fitted model with spatial
   component to test against targets.
}
