% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleGrid.R
\name{ScaleGrid}
\alias{ScaleGrid}
\title{Center and scale covariate list}
\usage{
ScaleGrid(Grid, vars = c("AvgSprSumPpt", "AvgSprSumTemp"), scale = TRUE)
}
\arguments{
\item{Grid}{A list of spatially gridded covariates.}

\item{vars}{A character vector naming the columns to scale.}
}
\value{
A list.
}
\description{
Center and scale covariate list
}
