% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDAST.R
\name{SimDAST}
\alias{SimDAST}
\title{Simulate the dynamic additive spatiotemporal (DAST) model.}
\usage{
SimDAST(theta, initialCover, eta, climate, years, colPr, colCover)
}
\arguments{
\item{theta}{A matrix of DAST parameters. Must be at least one row. Must
contain at least the following columns: `Beta[1]` (intercept),
`Beta[2]` (density-dependence), `Beta[3]` (precip effect),
`Beta[4]` (temperature effect), `sigma_y` (temporal standard deviation).}

\item{initialCover}{A matrix of initial cover values for each cell.}

\item{eta}{The empirical spatial offset for each cell. Must be same
dimensions as `initialCover`.}

\item{climate}{A matrix of future climate values. Rows are number of years
and columns are the precipitation (column 1) and temperature (column 2)
covariate values. Note that these typically need to be scaled
appropriately if the coefficients were standardized during model fitting.}

\item{years}{Number of years to simulate.}

\item{colPr}{The fitted probability of colonization if cover in a cell is 0.}

\item{colCover}{The average percent cover of colonized cells.}
}
\value{
A large array.
}
\description{
`SimDAST()` simulates the dynamic process defined by the DAST model, which
is built upon the Gompertz population model. The function allows for
prospective analyses of cover change under climate change scenarios,
for example, by the user supplying a `climate` matrix for future time
steps. The function returns a, typically large, array with dimensions
of: time, y (latitude), x (longitude), MCMC iterations. Note that a
colonization process is added to DAST for simulation to avoid permanent
extinction in cells that go to zero percent cover.
}
