% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TifToSf.R
\name{TifToSf}
\alias{TifToSf}
\title{Converts tif to sf object}
\usage{
TifToSf(coverPath, filename, resFactor = NULL)
}
\arguments{
\item{coverPath}{Character string of the path to cover data of interest.}

\item{filename}{A character string of the filename to read.}

\item{resFactor}{An integer for the factor by which to decrease
the resolution of the spatial data. Default is \code{NULL}, implying
no change to the native resolution.}
}
\value{
An sf object
}
\description{
Converts tif to sf object
}
