% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteNetCDF.R
\name{WriteNetCDF}
\alias{WriteNetCDF}
\title{Write the forecasts to compressed netCDF4 files.}
\usage{
WriteNetCDF(
  coreId,
  scenario,
  forecast_start_date,
  lons,
  lats,
  ensembles,
  outDir,
  forecast_flag,
  cover
)
}
\value{
Nothing
}
\description{
Write the forecasts to compressed netCDF4 files.
}
